+++ Tests/Functional/RenderingTest.php
@@ -33,7 +33,6 @@ class RenderingTest extends FunctionalTestCase {
    */
   protected $testExtensionsToLoad = [
     'typo3conf/ext/gridelements',
-    'typo3conf/ext/replacecontent',
     'typo3conf/ext/vhs',
     'typo3conf/ext/t3v_core',
     'typo3conf/ext/t3v_content'